+++ versioneer.py
@@ -1,4 +1,5 @@
 # mypy: ignore-errors
+# flake8: noqa
 # Version: 0.21
 
 """The Versioneer - like a rocketeer, but for versions.